It works exactly as I wrote
But not as well as I thought.

Things like `A B C + * D /` and `/ * A + B C D` eval differently even though
they are the same, just different notations.

Infix works perfectly however, infix works perfectly if evaluated from
left to right. So `A * ( B + C ) / D` doesn't work because I haven't made
parentheses functionality, but `B + C * A / D` works.

If I wanted to perfect this, I could search up parse trees. This isn't
the prepostinfix parser I dream of (at least not yet), but it's pretty
good for a days work, and what I'd use it for.