Only thing I needed to fix was formatting
Turns out this entire thing works off of a variable stack and a operator
queue. That is huge news, which means that I might even be able to do
this stuff with singly linked lists. Also it's just impressive to say I
did this with only a stack and a queue.

Anyways, I made a parser that solves both prefix, suffix and infix
problems (without precedence and parentheses). I'm a coding god.

P.S. Parentheses is easy. You can either do the messy thing and
recursively eval the thing inside of it, or have some sort of return
stack. Dang, it isn't over is it?